Finally Keen Dreams can be built as C++. Furthermore, with gcc v4.8.2 and certain warnings enabled, this codebase can be built as C or C++ with a total of zero warnings shown. Unfortunately id0_char_t had to be redefined as char (from uint8_t), otherwise a ton of casts like (const id0_byte_t *)"Some map" would be required. Hopefully a little fix to VWL_MeasureString, which was actually done for Catacomb 3-D (and presumably Keen 4-6 too) by id Software back in 1991, is all that's needed following this change. On a side-note, an unnecessary old copy of id_heads.h for cat3d is removed (unified cat3d/adventures header file is used instead).

+++ src/be_sdl.c
@@ -300,7 +300,7 @@ static void BEL_SDL_ParseConfig(void)
 			{
 				buffer[len-1] = '\0';
 			}
-			for (int i = 0; i < sizeof(g_sdlCfgEntries)/sizeof(*g_sdlCfgEntries); ++i)
+			for (int i = 0; i < (int)(sizeof(g_sdlCfgEntries)/sizeof(*g_sdlCfgEntries)); ++i)
 			{
 				if (!strncmp(g_sdlCfgEntries[i].cfgPrefix, buffer, strlen(g_sdlCfgEntries[i].cfgPrefix)))
 				{
